9주차_1차

+++ OpenGlSample/OpenGlSample/RenderableObject.h
@@ -6,7 +6,7 @@ class RenderableObject : public Object {
 public:
 	RenderableObject();
 
-	void SetTranslate(double x, double y, double z);
+	void SetTranslate(float x, float y, float z);
 	void IsPatrol(float x, float y, float z);
 	void SetObjectRange(double min_x, double max_x, double min_y, double max_y);
 	void SetAlpha(float a);
